p37-114
Switched add event button to events page and added filters
