+++ frontend/src/styles/App.css
@@ -58,4 +58,92 @@
     }
   }
   
+  .filters-container {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 10px;
+    margin: 20px 0;
+    padding: 10px;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  .filters-container select,
+  .filters-container button {
+    padding: 10px 14px;
+    font-size: 14px;
+    border-radius: 6px;
+    border: 1px solid #ddd;
+    background-color: #f8f9fa;
+    font-weight: 500;
+    transition: all 0.2s ease;
+  }
+  
+  .filters-container button {
+    background-color: #007bff;
+    color: #fff;
+    border: none;
+  }
+  
+  .filters-container .reset-btn {
+    background-color: #6c757d;
+  }
+  
+  .filters-container button:hover {
+    opacity: 0.9;
+    transform: scale(1.02);
+  }
+  
+  .event-form-container {
+    background-color: #ffffff;
+    padding: 20px;
+    border-radius: 12px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.05);
+    max-width: 800px;
+    margin: 20px auto;
+  }
+  
+  .event-form-container h3 {
+    font-size: 20px;
+    margin-bottom: 15px;
+    font-weight: 600;
+  }
+  
+  .form-group,
+  .form-group-row {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    margin-bottom: 15px;
+  }
+  
+  .form-group-row {
+    flex-direction: row;
+    justify-content: space-between;
+  }
+  
+  .form-group-row input,
+  .form-group-row select,
+  .event-form-container input,
+  .event-form-container select,
+  .event-form-container textarea {
+    flex: 1;
+    padding: 10px;
+    font-size: 14px;
+    border: 1px solid #ccc;
+    border-radius: 6px;
+  }
+  
+  .button-row {
+    display: flex;
+    gap: 10px;
+    justify-content: flex-start;
+  }
+  
+
+
+  .add-event-btn:hover,
+  .cancel-event-btn:hover {
+    opacity: 0.9;
+  }
